requirements.txt
shareithub
requests
colorama
python-dotenv
web3
py-solc-x 
pyfiglet
halo
